+++ genium/src/test/resources/smoke/maps/output/cpp/include/smoke/Maps.h
@@ -32,7 +32,7 @@ public:
 using ErrorCodeToMessageMap = ::std::unordered_map< int32_t, ::std::string >;
 using SomeId = ::std::string;
 using NumberToTypeDef = ::std::unordered_map< uint8_t, ::smoke::Maps::SomeId >;
-using TypeDefToNumber = ::std::unordered_map< ::smoke::Maps::SomeId, uint8_t, ::genium::hash< ::smoke::Maps::SomeId > >;
+using TypeDefToNumber = ::std::unordered_map< ::smoke::Maps::SomeId, uint8_t >;
 using ArrayOfInts = ::std::vector< int32_t >;
 using StringToArray = ::std::unordered_map< ::std::string, ::smoke::Maps::ArrayOfInts >;
 using SomeInt = int32_t;