+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Genium project Release Notes
 
+## Unreleased
+- Bug fixes:
+  + `std::hash` is now also used for typedefs to primitive types.
+
 ## 5.0.0
 - Release date: 2019-07-31
 - Features: